Move end-to-end tests to better place
