+++ WalletWasabi.Tests/UnitTests/WabiSabi/Integration/WabiSabiHttpApiIntegrationTests.cs
@@ -10,7 +10,7 @@ using WalletWasabi.WabiSabi.Backend.Models;
 using WalletWasabi.WabiSabi.Models.MultipartyTransaction;
 using Xunit;
 
-namespace WalletWasabi.Tests.RegressionTests.WabiSabi
+namespace WalletWasabi.Tests.UnitTests.WabiSabi.Integration
 {
 	public class WabiSabiHttpApiIntegrationTests : IClassFixture<WabiSabiApiApplicationFactory<Startup>>
 	{
